+++ components/Gnb/BadgeCounter/BadgeCounter.tsx
@@ -3,7 +3,7 @@ my dashboard에서 member들 보여주는 카운터
 우선 MockData와 연결
 */
 import { FC } from "react";
-import UserBadge from "@/components/Nav/UserBadge/UserBadge";
+import UserBadge from "@/components/Gnb/UserBadge/UserBadge";
 import { mockMembersData } from "@/components/Gnb/MockData";
 import useMaxDisplayCount from "@/hooks/useMaxDisplayCount";
